+++ R/RenouvTList.R
@@ -15,9 +15,12 @@
 ##'
 ##' @param threshold A numeric vector of thresholds.
 ##' 
-##' @param effDuration,distname.y Effective duration and distribution
-##'     See \code{\link{Renouv}}.
-##'
+##' @param effDuration,distname.y The effective duration and the
+##'     distribution name to be passed to \code{\link{Renouv}}. These
+##'     must be numeric/character vectors with length one. In other
+##'     words the same effective duration and the same distribution is
+##'     used for all thresholds.
+##' 
 ##' @param loopIni Logical. If \code{TRUE} the initial values of the
 ##'     parameters for a threshold are derived from the known
 ##'     converged cases obtained before. This is mainy useful in the
@@ -25,9 +28,26 @@
 ##'     problems should arise as long as no historical/censored
 ##'     observations are used. \bold{NOT IMPLEMENTED YET}.
 ##' 
+##' @param start.par.y Used to provide the parameter names and initial
+##'     values in \code{\link[Renext]{Renouv}} when the distribution
+##'     is not one of the special distributions. This can be either a
+##'     single vector of initial values to be used for all thresholds,
+##'     or a collection of initial values, one by threshold. In the
+##'     first case, \code{start.par.y} must be a named numeric vector
+##'     or a named list to be passed as the \code{start.par.y}
+##'     argument of \code{\link{Renouv}}. In the second case, one must
+##'     use either a numeric matrix with \code{length(threshold)}
+##'     rows, or a list of \code{length(threshold)} numeric vectors or
+##'     lists. The argument to be passed to \code{\link{Renouv}} will
+##'     then be given by looping over the rows of the matrix or over
+##'     the elements of the list.
+##'
 ##' @param ... Further arguments to be passed to
 ##'     \code{\link{Renouv}}. Graphical arguments should not be used.
 ##' 
+##' @param effDuration,distname.y Effective duration and distribution
+##'     See \code{\link{Renouv}}.
+##'
 ##' @return An object with (S3) class \code{"RenouvTList"}. This S3
 ##'     class has a number of methods.
 ##'
@@ -47,25 +67,66 @@
 ##'
 ##' methods(class = "RenouvTList")
 ##' autoplot(predict(fit))
-##' autoplot(coSd(fit))
+##' autoplot(coef(fit))
 ##'
 ##' ## Use more thresholds and compare with the jittered data
 ##' fit <- RenouvTList(Garonne,
 ##'                    threshold = seq(from = 2401, to = 3001, by = 10),
 ##'                    distname.y = "GPD")
-##' autoplot(coSd(fit))
+##' autoplot(coef(fit))
 ##' fitJit <- RenouvTList(GaronneJit,
 ##'                       threshold = seq(from = 2401, to = 3001, by = 10),
 ##'                       distname.y = "GPD")
-##' autoplot(coSd(fitJit))
+##' autoplot(coef(fitJit))
+##'
+##' ## Use an Extended GP Distribution fo the excesses. To get initial values
+##' ## for the parameters, we add a column of ones to the coefficients of
+##' ## 'fitJit'. Note that we must use `reParam = FALSE` to match the names
+##' ## of the parameters of the Renext GPD.
+##' fitJit0 <- RenouvTList(GaronneJit,
+##'                        threshold = seq(from = 2401, to = 3001, by = 50),
+##'                        distname.y = "GPD")
+##' spy <- cbind(kappa = 1.0, coef(fitJit0, reParam = FALSE))
+##' \dontrun{
+##' fitJit1 <- RenouvTList(GaronneJit,
+##'                        threshold = seq(from = 2401, to = 3001, by = 50),
+##'                        distname.y = "EGPD3", start.par.y = spy)
+##' autoplot(fitJit1)
+##' }
 RenouvTList <- function(x,
                         threshold = NULL,
                         effDuration = NULL,
                         distname.y = "GPD",
                         loopIni = FALSE,
+                        start.par.y = NULL,
                         ...) {
-    
+
     threshold <- sort(threshold)
+    
+    if (!is.null(spy <- start.par.y)) {
+        if (!is.null(u <- attr(spy, "threshold"))) {
+            if (!all.equal(u, threshold)) {
+                warning("'start.par.y' has a \"threshold\" attribute",
+                        "which differs from the given value of 'threshold'")
+            }
+        }
+        if (inherits(spy, "matrix")) {
+            if (nrow(spy) != length(threshold)) {
+                stop("if 'start.par.y' is given and is a matrix",
+                     " it must have `length(threshold)' rows")
+            }
+            spy <- lapply(seq_len(nrow(spy)), function(i) spy[i, ])
+        } else if (is.numeric(spy) || (is.list(spy) && !all(sapply(spy, is.list)))) {
+            spyL <- list()
+            for (iu in seq_along(threshold)) spyL[[iu]] <- spy
+            names(spyL) <- paste0("threshold = ", threshold) 
+            spy <- spyL
+        } else if (!is.list(spy) || length(spy) != length(threshold)) {
+            stop("Bad value for 'start.par.y'. See help for the accepted ",
+                 "types of values")
+        }
+    }
+    
 
     res <- list()
     
@@ -76,6 +137,7 @@ RenouvTList <- function(x,
                                effDuration = effDuration,
                                distname.y = distname.y,
                                plot = FALSE,
+                               start.par.y = spy[[i]],
                                ...)
             
         }
@@ -108,11 +170,21 @@ RenouvTList <- function(x,
 ##'     in the results. The choice \code{FALSE} is relevant for
 ##'     threshold stability analyzes.
 ##'
+##' @param sd Logical. If \code{TRUE} the result is a list of
+##'     \emph{two} matrices: one for the estimates and one for the
+##'     standard deviations of these, a.k.a. the \emph{standard
+##'     errors}.
+##' 
 ##' @param ... Not used yet.
 ##'
 ##' @return A numeric matrix with its rows corresponding to the
 ##'     thresholds.
 ##'
+##' @section Caution: the confidence intervals are obtained by using
+##'     the "delta" method. The intervals are known to have a coverage
+##'     rate which is smaller than the expected rate when the
+##'     Generalised Pareto Distribution is used.
+##' 
 ##' @export
 ##' @method coef RenouvTList
 ##'
@@ -122,7 +194,7 @@ RenouvTList <- function(x,
 ##'                    distname.y = "GPD")
 ##' autoplot(coef(fit))
 ##' 
-coef.RenouvTList <- function(object, reParam, lambda = TRUE, ...) {
+coef.RenouvTList <- function(object, reParam, lambda = TRUE, sd = TRUE, ...) {
     
     u <- attr(object, "threshold")
     dn <- sapply(object, function(x) x$distname.y)
@@ -135,25 +207,49 @@ coef.RenouvTList <- function(object, reParam, lambda = TRUE, ...) {
         reParam <- ifelse(dn %in% c("GPD"), TRUE, FALSE)
     }
     
-    mat <- t(sapply(object, coef))
+    est <- t(sapply(object, coef))
+    if (sd) sigma <- t(sapply(object, function(x) x$sigma))
+    
     if (!lambda) {
-        ind <- colnames(mat) != "lambda"
-        mat <- mat[ , ind]
+        ind <- colnames(est) != "lambda"
+        est <- est[ , ind]
+        if (sd) {
+            ind <- colnames(sigma) != "lambda"
+            sigma <- sigma[ , ind]
+        }
     }
     
     if (reParam) {
         if (dn %in% c("GPD")) {
-            mat[ , "scale"] <- mat[ , "scale"] - u * mat[ , "shape"]
-            colnames(mat) <- sub("^scale$", "scale ind", colnames(mat))
+            est[ , "scale"] <- est[ , "scale"] - u * est[ , "shape"]
+            colnames(est) <- sub("^scale$", "scale ind", colnames(est))
+
+            if (sd) {
+                Sigma <- lapply(object, vcov)
+                
+                for (i in seq_along(u)) {
+                    sigma[i, "scale"] <- sqrt(Sigma[[i]]["scale", "scale"] -
+                                              2 * u[i] * Sigma[[i]]["scale", "shape"] +
+                                              u[i]^2 * Sigma[[i]]["shape", "shape"])
+                }
+                colnames(sigma) <- sub("^scale$", "scale ind", colnames(est))
+            }
         } else {
             stop("'reParam' can only be 'TRUE' when the distribution ",
                  "is \"GPD\"")
         }
-    } 
+    }
+    if (sd) {
+        attr(est, "sd") <- TRUE
+        attr(est, "sigma") <- sigma
+    } else {
+        attr(est, "sd") <- FALSE
+    }
     
-    attr(mat, "threshold") <- attr(object, "threshold")
-    class(mat) <- "coef.RenouvTList"
-    mat
+    attr(est, "threshold") <- attr(object, "threshold")
+    class(est) <- "coef.RenouvTList"
+    est
+    
 }
 
 ## *****************************************************************************
@@ -163,9 +259,20 @@ coef.RenouvTList <- function(object, reParam, lambda = TRUE, ...) {
 ##' 
 print.coef.RenouvTList <- function(x, ...) {
     attr(x, "threshold") <- NULL
-    class(x) <- "matrix"
-    x <- round(x, digits = 3)
-    print(x)
+
+    if (!attr(x, "sd")) {
+        class(x) <- "matrix" 
+        x <- round(x, digits = 3)
+        attr(x, "sd") <- NULL
+        print(x)
+    } else {
+        mat <- array("", dim = dim(x), dimnames = dimnames(x))
+        for (i in 1:ncol(mat)) {
+            mat[ , i] <- sprintf("%6.3f [%5.3f]",
+                                 x[ , i], attr(x, "sigma")[ , i])
+        }
+        print(noquote(mat))
+    }
 }
 
 ## *****************************************************************************
@@ -212,18 +319,19 @@ print.coef.RenouvTList <- function(x, ...) {
 ##' @export
 coSd.RenouvTList <- function(object, reParam, lambda = TRUE,  ...)  {
 
+    .Deprecated("coef(object, sd = TRUE)")
+    
     u <- attr(object, "threshold")
     dn <- sapply(object, function(x) x$distname.y)
     if (length(dn <- unique(dn)) > 1) {
         stop("all elements of 'object' must have the same ",
              "'distname.y' element")
     }
-    
     if (missing(reParam)) {
-        reParam <- ifelse(dn  %in% c("GPD"), TRUE, FALSE)
+        reParam <- ifelse(dn %in% c("GPD"), TRUE, FALSE)
     }
-    est  <- t(sapply(object, coef))
-    sigma  <- t(sapply(object, function(x) x$sigma))
+    est <- t(sapply(object, coef))
+    sigma <- t(sapply(object, function(x) x$sigma))
     ## sigma  <- t(sapply(object, function(x) sqrt(diag(vcov(x)))))
     if (!lambda) {
         ind <- colnames(est) != "lambda"
@@ -239,7 +347,7 @@ coSd.RenouvTList <- function(object, reParam, lambda = TRUE,  ...)  {
             for (i in seq_along(u)) {
                 sigma[i, "scale"] <- sqrt(Sigma[[i]]["scale", "scale"] -
                                           2 * u[i] * Sigma[[i]]["scale", "shape"] +
-                    u[i]^2 *  Sigma[[i]]["shape", "shape"])
+                    u[i]^2 * Sigma[[i]]["shape", "shape"])
             }
             colnames(est) <- colnames(sigma) <- sub("^scale$", "scale ind", colnames(est))
         } else {
@@ -302,10 +410,15 @@ predict.RenouvTList <- function(object,
                                 ...) {
     
     noData <- (missing(newdata) || is.null(newdata))
-    if (noData) {
-        newdata <- as.vector(outer(c(1, 2, 3, 5, 7, 10.1), c(1, 10, 100)))
-    }
+   
+    ## The largest rate should be the first,  
+    lambdaMax <- max(coef(object)[ , "lambda"])
+    logGrid <- seq(from = -log(lambdaMax, base = 10) + 1e-6,
+                   to = log(1100, base = 10),
+                   length.out = 100)
     
+    newdata <- 10^logGrid
+
     predLong <- function(object, newdata, level, ...) {
         p <- predict(object, newdata = newdata, level = level, ...)
         names(p) <- c("Period", "Quantile", "L", "U")
@@ -328,7 +441,7 @@ predict.RenouvTList <- function(object,
 ##' 
 summary.RenouvTList <- function(object, ...) {
     x <- object
-    x$coSd <- coSd(object)
+    x$coSd <- coef(object, sd = TRUE)
     x$KS <- round(t(sapply(object,
                            function(o) c(n = o$nb.OT,
                                          o$KS$stat,
@@ -349,6 +462,6 @@ print.summary.RenouvTList <- function(x,
     cat("RenouvTList object\n")
     cat("o Estimated coefficients\n")
     print(x$coSd)
-    cat("o  Kolmogorov-Smirnov test\n")
+    cat("o Kolmogorov-Smirnov test\n")
     print(x$KS)
 }